mtg_card_pdf_generator/requirements.txt
requests
reportlab
pillow